app/requirements.txt
blinker==1.9.0
cffi==1.17.1
clarabel==0.11.1
click==8.2.1
colorama==0.4.6
cvxpy==1.7.0
ecdsa==0.19.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
mapping==0.1.6
MarkupSafe==3.0.2
numpy==2.3.1
osqp==1.0.4
pandas==2.3.1
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
scipy==1.16.0
scs==3.2.7.post2
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
